# api/swagger.yaml
...
swagger: "2.0"
info:
  description: "Hypermedia Applications project year 2018-2019. A bookstore with book\
    \ lists and user functions."
  version: "0.0.1"
  title: "Book House"
  termsOfService: ""
  license:
    name: "Apache 2.0"
    url: "http://www.apache.org/licenses/LICENSE-2.0.html"
host: "hypermedia-bookshop.herokuapp.com"
basePath: "/test"
...
schemes:
- "https"
- "http"
securityDefinitions:
  Bearer:
    type: apiKey
    name: Authorization
    in: header
paths:
...
  /authors:
    get:
      tags:
      - "author"
      summary: "Authors available in the database"
      description: "List of authors available in the database"
      operationId: "authorsGET"
      produces:
      - "application/json"
      parameters:
      - name: "offset"
        in: "query"
        description: "Pagination offset. Default is 0."
        required: false
        type: "integer"
      - name: "limit"
        in: "query"
        description: "Maximum number of items per page. Default is 20 and cannot exceed\
          \ 500."
        required: false
        type: "integer"
      responses:
        200:
          description: "A collection of Authors"
          schema:
            type: "array"
            items:
              $ref: "#/definitions/Author"
        404:
          description: "Authors not found"
      x-swagger-router-controller: "Author"
...
definitions:
...
  Author:
    type: "object"
    required:
    - "author_id"
    - "name"
    properties:
      author_id:
        type: "integer"
        format: "int64"
      name:
        type: "string"
      surname:
        type: "string"
      short_bio:
        type: "string"
      img_path:
        type: "string"
    title: "Author"
    description: "An author who writes wooks"
    example:
      author_id: "99"
      name: "Pedro"
      surname: "Infante"
      short_bio: "A great writer"
      img_path: "null"
...